--- test/file/too-big.spec.ts
@@ -1,8 +1,0 @@
-import { test } from '@playwright/test'
-import { createNote } from '../utils'
-import Files from './files'
-
-test.skip('to big zip', async ({ page }) => {
-  const files = [Files.Zip]
-  const link = await createNote(page, { files, error: 'note is to big' })
-})
